+++ tests/Database/Adapter/PdoAdapterTest.php
@@ -76,13 +76,24 @@ class PdoAdapterTest extends PHPUnit_Framework_TestCase
         $this->assertEquals(1, $this->adapter->insert('phoenix_test_table', ['id' => 1, 'title' => 'first', 'date' => new DateTime()]));
         $item = $this->adapter->fetch('phoenix_test_table', 'title', ['id' => 1]);
         $this->assertEquals('first', $item['title']);
+        
         $this->assertTrue($this->adapter->update('phoenix_test_table', ['id' => 1, 'title' => 'second', 'date' => new DateTime()], ['id' => 1]));
         $item = $this->adapter->fetch('phoenix_test_table', 'title', ['id' => 1], ['title']);
         $this->assertEquals('second', $item['title']);
+        
         $this->assertTrue($this->adapter->update('phoenix_test_table', ['id' => 1, 'title' => 'third', 'date' => new DateTime()], [], 'id = 1'));
         $items = $this->adapter->fetchAll('phoenix_test_table', 'title', ['id' => 1], null, ['title' => 'DESC'], ['id']);
         $this->assertEquals('third', $items[0]['title']);
 
+        $this->assertEquals(2, $this->adapter->insert('phoenix_test_table', ['id' => 2, 'title' => 'fourth', 'date' => new DateTime()]));
+        $this->assertTrue($this->adapter->update('phoenix_test_table', ['title' => 'multi update'], ['id' => [1, 2]]));
+        
+        $items = $this->adapter->fetchAll('phoenix_test_table', 'title', ['id' => [1, 2]]);
+        $this->assertCount(2, $items);
+        foreach ($items as $item) {
+            $this->assertEquals('multi update', $item['title']);
+        }
+        
         $this->setExpectedException('\Phoenix\Exception\DatabaseQueryExecuteException', 'SQLSTATE[HY000]: no such table: phoenix_non_exist_test_table.', 1);
         $this->adapter->update('phoenix_non_exist_test_table', ['id' => 1, 'title' => 'first', 'date' => new DateTime()]);
     }
@@ -155,6 +166,13 @@ class PdoAdapterTest extends PHPUnit_Framework_TestCase
         $this->adapter->delete('phoenix_test_table', [], 'id < 3');
         $items = $this->adapter->fetchAll('phoenix_test_table');
         $this->assertCount(0, $items);
+        
+        $this->assertEquals(1, $this->adapter->insert('phoenix_test_table', ['id' => 1, 'title' => 'first', 'date' => new DateTime()]));
+        $this->assertEquals(2, $this->adapter->insert('phoenix_test_table', ['id' => 2, 'title' => 'second', 'date' => new DateTime()]));
+        $items = $this->adapter->fetchAll('phoenix_test_table');
+        $this->assertCount(2, $items);
+        
+        $this->adapter->delete('phoenix_test_table', ['id' => [1, 2]]);
         
         $this->setExpectedException('\Phoenix\Exception\DatabaseQueryExecuteException', 'SQLSTATE[HY000]: no such table: phoenix_non_exist_test_table.', 1);
         $this->adapter->delete('phoenix_non_exist_test_table');