+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - support for change charset in mysql (table and also column)
 - simple changes of column types and names for pgsql
 - simple type casting for columns in pgsql
+- support for multi delete (IN condition)
 
 #### Fixed
 - several bugs in PdoAdapter